keyboard: split current and new shortcuts in edit dialog
When editing a shortcut, the current shortcut must be displayed
side-by-side with the new one. This gives a better understanding
of what's going on, and may avoid mistakes.

This patch splits the shortcut accelerator in 2: the current and
the new one, and also adapts the code to update only the new shortcut
label.

https://bugzilla.gnome.org/show_bug.cgi?id=769314