keyboard: make Add button insensitive after editing
When creating a new keyboard shortcut, the Add button
gets sensitive when all the fields are valid. If the
user tries to create a new shortcut right after closing,
the Add button is still sensitive even with invalid
custom fields.

Fix that by making the Add button sensitive whenever
we finish editing the new custom shortcut.

https://bugzilla.gnome.org/show_bug.cgi?id=769314